+++ app/utils/Filters.scala
@@ -1,9 +1,0 @@
-package utils
-
-import play.api.http.HttpFilters
-import play.filters.csrf.CSRFFilter
-import javax.inject.Inject
-
-class Filters @Inject() (csrfFilter: CSRFFilter) extends HttpFilters {
-  def filters = Seq(csrfFilter)
-}